Move SpatialVector in its own file
