+++ Sofa/framework/DefaultType/src/sofa/defaulttype/SolidTypes.inl
@@ -32,77 +32,6 @@ namespace sofa::defaulttype
 {
 
 template<class R>
-SolidTypes<R>::SpatialVector::SpatialVector( const Vec& l, const Vec& f ):lineVec(l),freeVec(f)
-{}
-
-template<class R>
-void SolidTypes<R>::SpatialVector::clear()
-{
-    lineVec = freeVec = Vec(0,0,0);
-}
-
-template<class R>
-typename SolidTypes<R>::SpatialVector& SolidTypes<R>::SpatialVector::operator += (const SpatialVector& v)
-{
-    lineVec += v.lineVec;
-    freeVec += v.freeVec;
-    return *this;
-}
-
-template<class R>
-typename SolidTypes<R>::SpatialVector SolidTypes<R>::SpatialVector::operator + ( const SpatialVector& v ) const
-{
-    return SpatialVector(lineVec+v.lineVec,freeVec+v.freeVec);
-}
-
-template<class R>
-typename SolidTypes<R>::SpatialVector SolidTypes<R>::SpatialVector::operator - ( const SpatialVector& v ) const
-{
-    return SpatialVector(lineVec-v.lineVec,freeVec-v.freeVec);
-}
-
-template<class R>
-typename SolidTypes<R>::SpatialVector SolidTypes<R>::SpatialVector::operator - ( ) const
-{
-    return SpatialVector(-lineVec,-freeVec);
-}
-
-/// Spatial dot product (cross terms)
-template<class R>
-typename SolidTypes<R>::Real SolidTypes<R>::SpatialVector::operator * ( const SpatialVector& v ) const
-{
-    return lineVec * v.freeVec + freeVec * v.lineVec;
-}
-
-/// Spatial cross product
-template<class R>
-typename SolidTypes<R>::SpatialVector SolidTypes<R>::SpatialVector::cross( const SpatialVector& v ) const
-{
-    return SpatialVector(
-            type::cross(lineVec,v.lineVec),
-            type::cross(freeVec,v.lineVec) + type::cross(lineVec,v.freeVec)
-            );
-}
-
-template<class R>
-typename SolidTypes<R>::SpatialVector SolidTypes<R>::SpatialVector::operator * (const Mat66& m) const
-{
-    SpatialVector result;
-    for( int i=0; i<3; i++ )
-    {
-        result.lineVec[i]=0;
-        result.freeVec[i]=0;
-        for( int j=0; j<3; j++ )
-        {
-            result.lineVec[i] += lineVec[j]*m[i][j] + freeVec[j]*m[i][j+3];
-            result.freeVec[i] += lineVec[j]*m[i+3][j] + freeVec[j]*m[i+3][j+3];
-        }
-    }
-    return result;
-}
-//======================================================================================================
-
-template<class R>
 SolidTypes<R>::Transform::Transform()
     : orientation_() // default constructor set to identity
     , origin_() // default constructor set to {0, 0, 0}
